refactor: PieceProperty의 인스턴스 Name을 계층으로 분류
