+++ src/main/java/chess/domain/piece/Pawn.java
@@ -3,7 +3,7 @@ package chess.domain.piece;
 import chess.domain.piece.property.Color;
 import chess.domain.piece.state.started.StartedPawn;
 
-public final class Pawn extends Piece {
+public class Pawn extends Piece {
     private static final String NAME = "p";
     private static final double SCORE = 1;
