+++ src/main/java/chess/domain/piece/Piece.java
@@ -9,12 +9,13 @@ import chess.domain.piece.property.Color;
 import chess.domain.piece.property.PieceProperty;
 import chess.domain.piece.state.PieceState;
 
-public abstract class Piece {
+public abstract class Piece extends Named{
     private final PieceProperty property;
     private PieceState state;
 
     public Piece(Color color, String name, double score, PieceState state) {
-        this.property = PieceProperty.of(color, name, score);
+        super(Name.of(name, color));
+        this.property = PieceProperty.of(color, score);
         this.state = state;
     }
 
@@ -37,10 +38,6 @@ public abstract class Piece {
         return property.isSameColor(piece.property);
     }
 
-    public String getName() {
-        return property.getName();
-    }
-
     public void updateState() {
         state = state.updateState();
     }
@@ -52,22 +49,5 @@ public abstract class Piece {
     public boolean isKing() {
         return new King(Color.White).isSame(this)
             || new King(Color.Black).isSame(this);
-    }
-
-    @Override
-    public boolean equals(Object o) {
-        if (this == o)
-            return true;
-        if (o == null || getClass() != o.getClass())
-            return false;
-
-        Piece piece = (Piece)o;
-
-        return getName() != null ? getName().equals(piece.getName()) : piece.getName() == null;
-    }
-
-    @Override
-    public int hashCode() {
-        return getName() != null ? getName().hashCode() : 0;
     }
 }
