+++ src/main/java/chess/domain/piece/property/Color.java
@@ -13,14 +13,6 @@ public enum Color {
         this.forward = forward;
     }
 
-    public String convertName(String name) {
-        if (this == Black) {
-            name = name.toUpperCase();
-        }
-
-        return name;
-    }
-
     public Direction forward() {
         return forward;
     }
