--- src/main/java/chess/domain/piece/property/Name.java
@@ -1,13 +1,0 @@
-package chess.domain.piece.property;
-
-public class Name {
-    private final String name;
-
-    public Name(String name) {
-        this.name = name;
-    }
-
-    public String getName() {
-        return name;
-    }
-}
